Add ReadLimit
This is the number of bytes that the receiver will eventually get
without further action from the sender (assuming the receiver doesn't
decide to close the connection).

Whenever the sender is blocked or idle, the receiver can read everything
in the buffer without further action from any other process.

ReceiverRead steps don't change ReadLimit, as long as the receiver
hasn't closed the connection. Therefore ReadLimit is correct:

- When the receiver is blocked it cannot read any more than it has
  without help.
- ReadLimit is Len(Got) then, so ReadLimit is obviously correct in this
  case.
- Since read steps preserve ReadLimit, this shows that ReadLimit is
  correct in all cases.

e.g. if ReadLimit = 5 and no other processes do anything, then we will
end up in a state with the receiver blocked, and ReadLimit = Len(Got) =
5 and so we really did read a total of 5 bytes.

ReadLimit never decreases (unless the receiver decides to close the connection).