WriteLimit
WriteLimit is the number of bytes that the sender will eventually send
without further action from the other processes or the client
application, assuming the connection isn't closed by either end.

Whenever the receiver is blocked, the sender will fill the buffer (or
write everything it wants to write) without further action from any
other process, according to WriteLimit.

Sender steps preserve WriteLimit (except for sender_ready, where the
application decides to send more data)

When in the ready or blocked states, WriteLimit predicts that we will
write exactly what we have already written, which is obviously the
correct answer. In other states, and if no other process runs, we must
eventually reach one of these known-correct states without performing a
sender_ready step. Therefore, WriteLimit will not have changed and it
must have been correct.

WriteLimit never decreases (unless the sender or receiver decides to
close the connection).

Also, constrain `free` to be zero in states where it is not live.
With this, we can remove the `Len(Buffer) = BufferSize` case from
`SenderInfoAccurate`.