Remove want variable and check optimised version
The C implementation needs to keep track of how big the application's
buffer is, but the algorithm only used it to set NotifyWrite if want >
have. However, that appears to be unnecessary, so remove want and allow
the algorithm to set NotifyWrite unnecessarily in any case.

This simplifies the invariants and proofs, and includes the behaviour of
the OCaml implementation (which doesn't read into a user-supplied
buffer) and the behaviour of QubesDB (which sets NotifyWrite in every
case).