Prove that WriteLimit is correct
If `WriteLimit` predicts that we'll eventually send the first `n` bytes,
then we will.

The main result is `WriterLive`, but there's also a proof of the weaker
`WriteLimitCorrect` that was there before. The proof would be simpler if
that were fixed to use `SenderFair`, but for now I've left it as it was.